Adding a new message functionality
Fixed Navbar redirecting "New Message" link to the wrong url.
Created a higher order component "withAuth" that makes sure a user is logged in before they can see the MessageForm container.
Created a MessageForm that allows a user to post a new message.
Created a postNewMessage action that creates a new message.
Added the "new message" Route to the Main container.